rename(ui-design-tool): rename TransformOverlay to SelectionOverlay
