+++ packages/ui-design-tool/components/SelectionOverlay/SelectionOverlay.css.ts
